requirements.txt
streamlit>=1.33,<2.0
python-dateutil>=2.9.0
